+++ src/main/java/com/oitsjustjose/naturalprogression/common/CommonProxy.java
@@ -3,6 +3,5 @@ package com.oitsjustjose.naturalprogression.common;
 import net.minecraft.entity.player.PlayerEntity;
 
 public class CommonProxy {
-    public void doHurtAnimation(PlayerEntity player) {
-    }
+    public void doHurtAnimation(PlayerEntity player) {}
 }