+++ src/main/java/com/oitsjustjose/naturalprogression/common/blocks/PebbleBlock.java
@@ -39,8 +39,9 @@ public class PebbleBlock extends Block implements IWaterLoggable {
     public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 
     public PebbleBlock() {
-        super(Properties.create(Material.EARTH, MaterialColor.LIGHT_GRAY).hardnessAndResistance(0.125F, 2F)
-                .sound(SoundType.STONE).doesNotBlockMovement().notSolid());
+        super(Properties.create(Material.EARTH, MaterialColor.LIGHT_GRAY)
+                .hardnessAndResistance(0.125F, 2F).sound(SoundType.STONE).doesNotBlockMovement()
+                .notSolid());
         this.setDefaultState(this.stateContainer.getBaseState().with(WATERLOGGED, Boolean.FALSE));
     }
 
@@ -54,9 +55,11 @@ public class PebbleBlock extends Block implements IWaterLoggable {
 
     @Override
     @Nonnull
-    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
+    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos,
+            ISelectionContext context) {
         Vector3d offset = state.getOffset(worldIn, pos);
-        return VoxelShapes.create(0.37D, 0.0D, 0.37D, 0.69D, 0.065D, 0.69D).withOffset(offset.x, offset.y, offset.z);
+        return VoxelShapes.create(0.37D, 0.0D, 0.37D, 0.69D, 0.065D, 0.69D).withOffset(offset.x,
+                offset.y, offset.z);
     }
 
     @Override
@@ -72,8 +75,8 @@ public class PebbleBlock extends Block implements IWaterLoggable {
     }
 
     @Override
-    public ActionResultType onBlockActivated(BlockState state, World worldIn, BlockPos pos, PlayerEntity player,
-            Hand handIn, BlockRayTraceResult hit) {
+    public ActionResultType onBlockActivated(BlockState state, World worldIn, BlockPos pos,
+            PlayerEntity player, Hand handIn, BlockRayTraceResult hit) {
         if (!player.isCrouching()) {
             worldIn.destroyBlock(pos, true);
             player.swingArm(handIn);
@@ -102,20 +105,22 @@ public class PebbleBlock extends Block implements IWaterLoggable {
     @Nonnull
     @SuppressWarnings("deprecation")
     public FluidState getFluidState(BlockState state) {
-        return state.get(WATERLOGGED) ? Fluids.WATER.getStillFluidState(false) : super.getFluidState(state);
+        return state.get(WATERLOGGED) ? Fluids.WATER.getStillFluidState(false)
+                : super.getFluidState(state);
     }
 
     @Override
     @SuppressWarnings("deprecation")
-    public void neighborChanged(BlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos,
-            boolean isMoving) {
+    public void neighborChanged(BlockState state, World worldIn, BlockPos pos, Block blockIn,
+            BlockPos fromPos, boolean isMoving) {
         super.neighborChanged(state, worldIn, pos, blockIn, fromPos, isMoving);
         if (!this.isValidPosition(state, worldIn, pos)) {
             worldIn.destroyBlock(pos, true);
         }
         // Update the water from flowing to still or vice-versa
         else if (state.get(WATERLOGGED)) {
-            worldIn.getPendingFluidTicks().scheduleTick(pos, Fluids.WATER, Fluids.WATER.getTickRate(worldIn));
+            worldIn.getPendingFluidTicks().scheduleTick(pos, Fluids.WATER,
+                    Fluids.WATER.getTickRate(worldIn));
         }
     }