+++ src/main/java/com/oitsjustjose/naturalprogression/common/items/HatchetItem.java
@@ -8,7 +8,9 @@ import net.minecraftforge.common.ToolType;
 
 public class HatchetItem extends AxeItem {
     public HatchetItem() {
-        super(NaturalProgressionItems.flintTier, 1.8F, 0F, new Item.Properties().addToolType(ToolType.AXE, 0)
-                .group(NaturalProgressionGroup.getInstance()).maxStackSize(1).maxDamage(16));
+        super(NaturalProgressionItems.flintTier, 1.8F, 0F,
+                new Item.Properties().addToolType(ToolType.AXE, 0)
+                        .group(NaturalProgressionGroup.getInstance()).maxStackSize(1)
+                        .maxDamage(16));
     }
 }